ai: use a safer way of listing files
pathlib's glob can fail on WinFsp mountpoints when it attempts to
determine whether a path is a directory, which is problematic
given that content roots will most likely be FUSE mountpoints.

This commit wraps the glob call inside of a try-catch block to
add a simple fallback in case an error occurs and prevent crashes.